refactor the reading for varOne configuration
